3f5a9a7kq: integrated email notification
