--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "nodemailer": "^6.9.14"
-  }
-}
