+++ server/src/services/registerstudent.ts
@@ -2,6 +2,7 @@
 
 import prisma from '@/lib/prisma';
 import bcrypt from 'bcrypt';
+import { sendEmail } from '@/app/api/email/route';
 
 interface RegisterStudentData {
   firstName: string;
@@ -11,7 +12,7 @@ interface RegisterStudentData {
 
 export async function registerStudent(data: RegisterStudentData) {
   const { firstName, lastName, email } = data;
-  const password = Math.random().toString(36).slice(-8); // Generate a random password
+  const password = Math.random().toString(36).slice(-8); 
 
   const hashedPassword = await bcrypt.hash(password, 10);
 
@@ -24,6 +25,12 @@ export async function registerStudent(data: RegisterStudentData) {
         passwordHash: hashedPassword,
         role: 'student',
       },
+    });
+
+    await sendEmail({
+      email,
+      name: `${firstName} ${lastName}`,
+      message: `Your registration was successful. Your temporary password is: ${password}`,
     });
   } catch (error) {
     console.error('Error registering student:', error);
